requirements.txt
numpy>=1.26.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
pycairo>=1.24.0
PyGObject
pytest
